+++ app/src/main/jni/mupen64plus-video-rice/projects/unix/Makefile
@@ -339,7 +339,8 @@ SOURCE = \
 	$(SRCDIR)/TextureFilters_hq4x.cpp \
 	$(SRCDIR)/TextureManager.cpp \
 	$(SRCDIR)/VectorMath.cpp \
-	$(SRCDIR)/Video.cpp
+	$(SRCDIR)/Video.cpp \
+	$(SRCDIR)/VR.cpp
 
 ifeq ($(OS),MINGW)
 SOURCE += \
